requirements.txt
fastapi
uvicorn
gunicorn
tensorflow
pillow
python-multipart
